Move Dashboard View into dashboards package
