+++ ui/dashboards/src/views/index.ts
@@ -11,6 +11,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export * from './components';
-export * from './context';
-export * from './views';
+export * from './ViewDashboard';
